chore(ci): Try a fixed version of create-pull-request
The update to actions/checkout v6 breaks create-pull-request because it
moves the configuration of the Authorization header to an included file
instead of where create-pull-request knows to override it.

Temporarily use a version with a purported fix to validate whether the
fix actually works for the case where a different token is used to
create the PR than was used to check out the repo.

Signed-off-by: Erik Swanson <erik@retailnext.net>